+++ sdk/identity/identity/src/msal/nodeFlows/msalClient.ts
@@ -72,36 +72,22 @@ export interface GetTokenInteractiveOptions extends GetTokenWithSilentAuthOption
  */
 export interface MsalClient {
   /**
-   * Retrieves an access token by using the on-behalf-of flow and a client certificate of the calling service.
+   *
+   * Retrieves an access token by using the on-behalf-of flow and a client assertion callback of the calling service.
    *
    * @param scopes - The scopes for which the access token is requested. These represent the resources that the application wants to access.
    * @param userAssertionToken - The access token that was sent to the middle-tier API. This token must have an audience of the app making this OBO request.
-   * @param clientCertificate - The client certificate used for authentication.
+   * @param clientCredentials - The client secret OR client certificate OR client `getAssertion` callback.
    * @param options - Additional options that may be provided to the method.
    * @returns An access token.
    */
   getTokenOnBehalfOf(
     scopes: string[],
     userAssertionToken: string,
-    clientCertificate: CertificateParts,
+    clientCredentials: string | CertificateParts | (() => Promise<string>),
     options?: GetTokenOptions,
   ): Promise<AccessToken>;
-  /**
-   *
-   * Retrieves an access token by using the on-behalf-of flow and a client secret of the calling service.
-   *
-   * @param scopes - The scopes for which the access token is requested. These represent the resources that the application wants to access.
-   * @param userAssertionToken - The access token that was sent to the middle-tier API. This token must have an audience of the app making this OBO request.
-   * @param clientSecret - The client secret used for authentication.
-   * @param options - Additional options that may be provided to the method.
-   * @returns An access token.
-   */
-  getTokenOnBehalfOf(
-    scopes: string[],
-    userAssertionToken: string,
-    clientSecret: string,
-    options?: GetTokenOptions,
-  ): Promise<AccessToken>;
+
   /**
    * Retrieves an access token by using an interactive prompt (InteractiveBrowserCredential).
    * @param scopes - The scopes for which the access token is requested. These represent the resources that the application wants to access.
@@ -158,13 +144,13 @@ export interface MsalClient {
    * Retrieves an access token by using a client assertion.
    *
    * @param scopes - The scopes for which the access token is requested. These represent the resources that the application wants to access.
-   * @param clientAssertion - The client assertion used for authentication.
+   * @param clientAssertion - The client `getAssertion` callback used for authentication.
    * @param options - Additional options that may be provided to the method.
    * @returns An access token.
    */
   getTokenByClientAssertion(
     scopes: string[],
-    clientAssertion: string,
+    clientAssertion: () => Promise<string>,
     options?: GetTokenOptions,
   ): Promise<AccessToken>;
 
@@ -405,7 +391,9 @@ export function createMsalClient(
 
     // Initialize a new app and cache it
     state.logger.getToken.info(
-      `Creating new ConfidentialClientApplication with CAE ${options.enableCae ? "enabled" : "disabled"}.`,
+      `Creating new ConfidentialClientApplication with CAE ${
+        options.enableCae ? "enabled" : "disabled"
+      }.`,
     );
 
     const cachePlugin = options.enableCae
@@ -563,7 +551,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
 
   async function getTokenByClientAssertion(
     scopes: string[],
-    clientAssertion: string,
+    clientAssertion: () => Promise<string>,
     options: GetTokenOptions = {},
   ): Promise<AccessToken> {
     state.logger.getToken.info(`Attempting to acquire token using client assertion`);
@@ -711,34 +699,26 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
     });
   }
 
-  function getTokenOnBehalfOf(
-    scopes: string[],
-    userAssertionToken: string,
-    clientSecret: string,
-    options?: GetTokenOptions,
-  ): Promise<AccessToken>;
-  function getTokenOnBehalfOf(
-    scopes: string[],
-    userAssertionToken: string,
-    clientCertificate: CertificateParts,
-    options?: GetTokenOptions,
-  ): Promise<AccessToken>;
   async function getTokenOnBehalfOf(
     scopes: string[],
     userAssertionToken: string,
-    clientSecretOrCertificate: string | CertificateParts,
+    clientCredentials: string | CertificateParts | (() => Promise<string>),
     options: GetTokenOptions = {},
   ): Promise<AccessToken> {
     msalLogger.getToken.info(`Attempting to acquire token on behalf of another user`);
 
-    if (typeof clientSecretOrCertificate === "string") {
+    if (typeof clientCredentials === "string") {
       // Client secret
       msalLogger.getToken.info(`Using client secret for on behalf of flow`);
-      state.msalConfig.auth.clientSecret = clientSecretOrCertificate;
+      state.msalConfig.auth.clientSecret = clientCredentials;
+    } else if (typeof clientCredentials === "function") {
+      // Client Assertion
+      msalLogger.getToken.info(`Using client assertion callback for on behalf of flow`);
+      state.msalConfig.auth.clientAssertion = clientCredentials;
     } else {
       // Client certificate
       msalLogger.getToken.info(`Using client certificate for on behalf of flow`);
-      state.msalConfig.auth.clientCertificate = clientSecretOrCertificate;
+      state.msalConfig.auth.clientCertificate = clientCredentials;
     }
 
     const msalApp = await getConfidentialApp(options);